split operations.py into resources
