+++ pyblaise/operations/exceptions.py
@@ -1,0 +1,5 @@
+class CreateRoleFailed(Exception):
+    """
+    raised when a role cannot be created in a server park
+    """
+    pass
